homeworks/hw-1/requirements.txt
torch>=2.2.0
transformers>=4.39.0
datasets>=2.19.0
accelerate>=0.30.0
tqdm>=4.66.0
Pillow>=10.1.0
numpy>=1.26.0
pandas>=2.2.0
annoy>=1.17.3
gradio>=4.29.0
huggingface_hub>=0.22.0
safetensors>=0.4.2
matplotlib>=3.8.0
plotly>=5.20.0
scikit-learn>=1.4.0
